Reorganize developer guide.
The main change is that all the "guides" are now under the "guidelines" session.
This reorganization left a few pages (including the front page) empty.
We should fill these up with content later but I am trying to avoid makind changes and reorganize the page at the same time to make it easier to review.